refactor: renamed selectors, typed
