+++ src/app/todo-list/todo-card/todo-card.component.ts
@@ -1,25 +1,21 @@
-import { Component, EventEmitter, Input, Output } from "@angular/core";
-
-
+import { Component, EventEmitter, Input, Output } from '@angular/core';
+import { Todo } from '../todo-list.component';
 
 @Component({
-    selector: 'todo-card-root',
-    templateUrl: './todo-card.component.html',
-    styleUrl: './todo-card.component.scss',
-    standalone: true,
-    imports: []
+  selector: 'app-todo-card',
+  templateUrl: './todo-card.component.html',
+  styleUrl: './todo-card.component.scss',
+  standalone: true,
+  imports: [],
 })
-
 export class TodoCardComponent {
-    @Input()
+  @Input()
+  todoCard!: Todo;
 
-    todoCard: any
+  @Output()
+  deleteTodoCard = new EventEmitter();
 
-    @Output()
-
-    deleteTodoCard = new EventEmitter()
-
-    onDeleteTodo(todoCardId: number) {
-        this.deleteTodoCard.emit(todoCardId)
-    }
+  onDeleteTodo(todoCardId: number) {
+    this.deleteTodoCard.emit(todoCardId);
+  }
 }
