+++ src/app/users-list/user-card/user-card.component.html
@@ -1,5 +1,5 @@
 <p>{{ userCard.name }}</p>
-<p>{{ userCard.company.name }}</p>
+<p>{{ userCard.company?.name }}</p>
 <p>{{ userCard.email }}</p>
 <p>{{ userCard.website }}</p>
 <button (click)="onDeleteUser(userCard.id)">Удалить карточку</button>
