+++ src/app/users-list/users-list.component.html
@@ -1,6 +1,6 @@
 <app-user-form (createUser)="createUser($event)"></app-user-form>
 <div class="users_list_cont">
   <div class="user" *ngFor="let user of usersService.usersSubject$ | async">
-   <user-card-root [userCard]="user" (deleteUserCard)="deleteUser($event)"></user-card-root>
+   <app-user-card [userCard]="user" (deleteUserCard)="deleteUser($event)"></app-user-card>
   </div>
 </div>
